# front/docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Stocky API
  version: 0.1.0
  description: API REST do sistema Stocky.
servers:
  - url: http://localhost:3000
    description: Ambiente local do back-end
...
paths:
  /auth/me:
    post:
      tags:
        - Auth
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Verificar usuario autenticado
      description: Verifica se o token informado ainda e valido e retorna o usuario autenticado.
      operationId: verificarAutenticacao
      responses:
        "200":
          description: Usuario autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
              examples:
                autenticado:
                  value:
                    id: 1
                    nome: Maria Silva
                    cpf: "12345678901"
                    email: maria.silva@example.com
                    ativo: true
        "401":
          description: Autenticacao nao realizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                autenticacaoNaoRealizada:
                  value:
                    error: Autenticação não realizada
        "500":
          description: Erro inesperado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                erroInterno:
                  value:
                    error: Erro interno do servidor
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token informado no header Authorization.
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Token informado no cookie httpOnly `token`.
  schemas:
...
    Usuario:
      type: object
      required:
        - id
        - nome
        - cpf
        - email
        - ativo
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Maria Silva
        cpf:
          type: string
          minLength: 11
          maxLength: 11
          example: "12345678901"
        email:
          type: string
          format: email
          example: maria.silva@example.com
        ativo:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Campos obrigatorios ausentes
        details:
          type: object
          additionalProperties: true